+++ template/packages/frontend/components/GlobalStyles.ts
