+++ template/packages/frontend/components/IndexStyles.ts
